CameraController: Clamp target values to within camera world
Ok, this is a really small thing but let me make my case. Now, targets
get clamped so that the camera never tries to look past its world
bounds.

This means the camera pans smoothly even when you go from one side of
the map to the other (instead of shooting past it and getting blocked
from continuing).